Initial commit
The code currently exists in a poorly documented and structured manner as
it is really more a proof of concept at this point. It supports installing
through pip and being run as a module. As of now pugs can be created,
deleted, and players can join and leave roles and pugs.